Completed Submission
Work was done on my laptop which unfortunately I did not sync with github, which is why this commit contains many changes that took much time. All the changes cannot be summarized here but will be in final report. Schematic of physical circuit made using LTSpice has also been added.